unsigned long long → uint64_t
For portability

+++ src/int_stuff.cpp
@@ -6,12 +6,12 @@
 #ifdef _WIN32
 	std::ostream &operator<<(std::ostream &strm, unsigned __uint128 value)
 	{
-		const unsigned long long uLongLongMax = 18446744073709551615ULL;
-		const unsigned long long uNineteenZeroes = 10000000000000000000ULL;
+		const uint64_t uLongLongMax = 18446744073709551615ULL;
+		const uint64_t uNineteenZeroes = 10000000000000000000ULL;
 		if (value > uLongLongMax)
 		{
 			__uint128 first = value / uNineteenZeroes;
-			unsigned long long rest = value % uNineteenZeroes;
+			uint64_t rest = value % uNineteenZeroes;
 			operator<<(strm, first);
 			strm << std::setfill ('0') << std::setw (19);
 			strm << rest;
@@ -19,7 +19,7 @@
 		else
 		{
 			// Printable directly
-			unsigned long long shorterValue = value;
+			uint64_t shorterValue = value;
 			strm << shorterValue;
 		}
 		return strm;
@@ -28,8 +28,8 @@
 	std::ostream &operator<<(std::ostream &strm, __int128 value)
 	{
 		strm << "value: " << value << std::endl;
-		const unsigned long long uLongLongMax = 18446744073709551615ULL;
-		const unsigned long long uNineteenZeroes = 10000000000000000000ULL;
+		const uint64_t uLongLongMax = 18446744073709551615ULL;
+		const uint64_t uNineteenZeroes = 10000000000000000000ULL;
 		if (value < 0)
 		{
 			strm << "(negative value)";
@@ -37,7 +37,7 @@
 		else if (value > uLongLongMax)
 		{
 			__int128 first = value / uNineteenZeroes;
-			unsigned long long rest = value % uNineteenZeroes;
+			uint64_t rest = value % uNineteenZeroes;
 			operator<<(strm, first);
 			strm << std::setfill ('0') << std::setw (19);
 			strm << rest;
@@ -45,7 +45,7 @@
 		else
 		{
 			// Printable directly
-			unsigned long long shorterValue = value;
+			uint64_t shorterValue = value;
 			strm << shorterValue;
 		}
 		return strm;
@@ -54,12 +54,12 @@
 	std::ostream &operator<<(std::ostream &strm, __uint128_t value)
 	{
 		strm << "value2: " << value << std::endl;
-		const unsigned long long uLongLongMax = 18446744073709551615ULL;
-		const unsigned long long uNineteenZeroes = 10000000000000000000ULL;
+		const uint64_t uLongLongMax = 18446744073709551615ULL;
+		const uint64_t uNineteenZeroes = 10000000000000000000ULL;
 		if (value > uLongLongMax)
 		{
 			__uint128_t first = value / uNineteenZeroes;
-			unsigned long long rest = value % uNineteenZeroes;
+			uint64_t rest = value % uNineteenZeroes;
 			operator<<(strm, first);
 			strm << std::setfill ('0') << std::setw (19);
 			strm << rest;
@@ -67,7 +67,7 @@
 		else
 		{
 			// Printable directly
-			unsigned long long shorterValue = value;
+			uint64_t shorterValue = value;
 			strm << shorterValue;
 		}
 		return strm;
@@ -75,8 +75,8 @@
 	
 	std::ostream &operator<<(std::ostream &strm, __int128_t value)
 	{
-		const unsigned long long uLongLongMax = 18446744073709551615ULL;
-		const unsigned long long uNineteenZeroes = 10000000000000000000ULL;
+		const uint64_t uLongLongMax = 18446744073709551615ULL;
+		const uint64_t uNineteenZeroes = 10000000000000000000ULL;
 		if (value < 0)
 		{
 			strm << "(negative value)";
@@ -84,7 +84,7 @@
 		else if (value > uLongLongMax)
 		{
 			__int128_t first = value / uNineteenZeroes;
-			unsigned long long rest = value % uNineteenZeroes;
+			uint64_t rest = value % uNineteenZeroes;
 			operator<<(strm, first);
 			strm << std::setfill ('0') << std::setw (19);
 			strm << rest;
@@ -92,7 +92,7 @@
 		else
 		{
 			// Printable directly
-			unsigned long long shorterValue = value;
+			uint64_t shorterValue = value;
 			strm << shorterValue;
 		}
 		return strm;
